requirements.txt
# W119 Capstone — Pinned Dependencies
# Python 3.12.6
# Install: pip install -r requirements.txt

# Core data
pandas==2.2.3
numpy==1.26.4
pyreadstat==1.3.3

# Modeling
scikit-learn==1.5.2
scipy==1.13.1
statsmodels==0.14.4

# Visualization
matplotlib==3.10.1
seaborn==0.13.2

# Interpretability
shap==0.48.0

# Config
PyYAML==6.0.2

# Notebooks
jupyter>=1.0
ipykernel>=6.0

# Latent modeling (Phase 7)
factor-analyzer==0.5.1
